Create one file per system
